+++ src/backend/joanie/payment/models.py
@@ -356,7 +356,7 @@ class CreditCardManager(models.Manager):
         payment_provider = get_payment_backend()
 
         return self.get(
-            pk=pk, owner__username=username, payment_provider=payment_provider.name
+            pk=pk, owners__username=username, payment_provider=payment_provider.name
         )
 
     def get_cards_for_owner(self, username):
@@ -367,8 +367,67 @@ class CreditCardManager(models.Manager):
         payment_provider = get_payment_backend()
 
         return self.filter(
-            owner__username=username, payment_provider=payment_provider.name
+            owners__username=username, payment_provider=payment_provider.name
         )
+
+
+class CreditCardOwnership(BaseModel):
+    """
+    CreditCardOwnership model allows to define the ownership of a user and a credit card,
+    finally it defines when it's the main one.
+    """
+
+    owner = models.ForeignKey(
+        to=User, on_delete=models.CASCADE, related_name="credit_card_ownerships"
+    )
+    credit_card = models.ForeignKey(
+        to="payment.CreditCard", on_delete=models.CASCADE, related_name="ownerships"
+    )
+    is_main = models.BooleanField(_("main"), default=False)
+
+    class Meta:
+        db_table = "joanie_credit_card_ownership"
+        verbose_name = "credit_card_ownership"
+        verbose_name_plural = "credit card ownernships"
+        ordering = ["-created_on"]
+        constraints = [
+            models.UniqueConstraint(
+                fields=["owner", "is_main"],
+                condition=models.Q(is_main=True),
+                name="unique_main_credit_card_per_user",
+            ),
+            models.UniqueConstraint(
+                fields=["credit_card", "owner"], name="unique_credit_card_ownership"
+            ),
+        ]
+
+    def clean(self, *args, **kwargs):
+        """
+        First if this is the user's first credit card, we enforce `is_main` to True.
+        Else if we are promoting a credit card as main, we demote the existing main credit card
+        Finally we prevent to demote the main credit card directly.
+        """
+        if not self.__class__.objects.filter(owner=self.owner).exists():
+            self.is_main = True
+        elif self.is_main:
+            self.__class__.objects.filter(owner=self.owner, is_main=True).update(
+                is_main=False
+            )
+        elif (
+            self.created_on
+            and not self.is_main
+            and self.__class__.objects.filter(
+                owner=self.owner, credit_card=self.credit_card, is_main=True
+            ).exists()
+        ):
+            raise ValidationError(_("Demote a main credit card is forbidden"))
+
+        return super().clean()
+
+    def save(self, *args, **kwargs):
+        """Enforce validation each time an instance is saved."""
+        self.full_clean()
+        super().save(*args, **kwargs)
 
 
 class CreditCard(BaseModel):
@@ -399,13 +458,23 @@ class CreditCard(BaseModel):
     )
     expiration_year = models.PositiveSmallIntegerField(_("expiration year"))
     last_numbers = models.CharField(_("last 4 numbers"), max_length=4)
+    # Deprecated
     owner = models.ForeignKey(
         to=User,
         verbose_name=_("owner"),
         related_name="credit_cards",
         on_delete=models.CASCADE,
+        blank=True,
+        null=True,
     )
-    is_main = models.BooleanField(_("main"), default=False)
+    owners = models.ManyToManyField(
+        to=User,
+        verbose_name=_("owners"),
+        related_name="payment_cards",
+        through="CreditCardOwnership",
+        through_fields=("credit_card", "owner"),
+    )
+    is_main = models.BooleanField(_("main"), default=False)  # Deprecated
     payment_provider = models.CharField(
         _("payment provider"), max_length=50, null=True, blank=True
     )
@@ -415,33 +484,25 @@ class CreditCard(BaseModel):
         verbose_name = "credit card"
         verbose_name_plural = "credit cards"
         ordering = ["-created_on"]
-        constraints = [
-            models.UniqueConstraint(
-                condition=models.Q(is_main=True),
-                fields=["owner"],
-                name="unique_main_credit_card_per_user",
-            )
-        ]
+
+    def add_owner(self, owner):
+        """
+        Add new owner to the credit card owners field through the `CreditCardOwnership` model to
+        enforce the validation logic (through the clean() method) for the field `is_main`
+        of ownership each time it's saved.
+        """
+        CreditCardOwnership.objects.create(
+            owner=owner,
+            credit_card=self,
+        )
 
     def clean(self):
         """
-        First if this is the user's first credit card, we enforce is_main to True.
-        Else if we are promoting an credit card as main, we demote the existing main credit card
-        Finally prevent to demote the main credit card directly.
+        It's required to have a `payment_provider` value, because we add credit card through a
+        payment provider only.
         """
         if not self.payment_provider:
             raise ValidationError(_("Payment provider field cannot be None."))
-        if not self.owner.credit_cards.exists():
-            self.is_main = True
-        elif self.is_main is True:
-            self.owner.credit_cards.filter(is_main=True).update(is_main=False)
-        elif (
-            self.created_on
-            and self.is_main is False
-            and self.owner.credit_cards.filter(is_main=True, pk=self.pk).exists()
-        ):
-            raise ValidationError(_("Demote a main credit card is forbidden"))
-
         return super().clean()
 
     def save(self, *args, **kwargs):